GX3-2958 Creado pakage con el objeto trello para reutilización de codigo y mayor legibilidad del código con funciones:
